member service repository 삭제
