--- WeverseEgg0.2/src/main/java/com/java/service/member/MemberService.java
@@ -1,18 +1,0 @@
-package com.java.service.member;
-
-import java.util.Optional;
-
-import com.java.dto.member.MemberDto;
-import com.java.entity.member.MemberEntity;
-
-public interface MemberService {
-
-	// singin 시작 시 이메일 입력 
-	MemberDto findByEmail(String email);
-
-	boolean createAccount(String email, String pw, String nickname);
-
-	MemberDto findByEmailAndPw(String email, String pw);
-
-
-}
